+++ test/spmd/tensor/test_tp_sharding_ops.py
@@ -38,126 +38,6 @@ class TPShardingOpsTest(DistTensorTestBase):
         self.assertTrue(new_dt.placements[0].is_shard(dim=0))
         self.assertEqual(new_dt.to_local(), tensor.transpose(1, 2))
 
-    # TODO: Need to investigate why test failed in CPU for baddbmm.
-    @with_comms
-    @skip_if_lt_x_gpu(TEST_GPU_NUM)
-    def test_sharded_baddbmm(self):
-        # If beta is 0, input tensor will be ignored
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        torch.manual_seed(self.rank)
-        tensor = torch.rand(3, 5, 6, device=self.device_type)
-        batch_1 = torch.rand(3, 5, 8, device=self.device_type)
-        batch_2 = torch.rand(3, 8, 6, device=self.device_type)
-        local_result = torch.baddbmm(
-            tensor, batch_1, batch_2, beta=0.0, alpha=0.5
-        )
-        sharding = [Shard(0)]
-        tensor_dt = DTensor.from_local(tensor, device_mesh, sharding)
-        batch_1_dt = DTensor.from_local(batch_1, device_mesh, sharding)
-        batch_2_dt = DTensor.from_local(batch_2, device_mesh, sharding)
-        new_dt = torch.baddbmm(
-            tensor_dt, batch_1_dt, batch_2_dt, beta=0.0, alpha=0.5
-        )
-        self.assertTrue(new_dt.placements[0].is_shard(dim=0))
-        self.assertEqual(new_dt.to_local(), local_result)
-
-        # If beta != 0
-        local_result = torch.baddbmm(
-            tensor, batch_1, batch_2, beta=0.8, alpha=0.5
-        )
-        new_dt = torch.baddbmm(
-            tensor_dt, batch_1_dt, batch_2_dt, beta=0.8, alpha=0.5
-        )
-        self.assertTrue(new_dt.placements[0].is_shard(dim=0))
-        self.assertEqual(new_dt.to_local(), local_result)
-
-    @with_comms
-    @skip_if_lt_x_gpu(TEST_GPU_NUM)
-    def test_sharded_baddbmm_with_bwd(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        torch.manual_seed(10)
-        tensor = torch.rand(
-            8, 5, 6, device=self.device_type, requires_grad=True
-        )
-        batch_1 = torch.rand(
-            8, 5, 8, device=self.device_type, requires_grad=True
-        )
-        batch_2 = torch.rand(
-            8, 8, 6, device=self.device_type, requires_grad=True
-        )
-        local_result = torch.baddbmm(
-            tensor, batch_1, batch_2, beta=0.0, alpha=0.5
-        )
-        sharding = [Shard(0)]
-        tensor_dt = distribute_tensor(tensor, device_mesh, sharding)
-        batch_1_dt = distribute_tensor(batch_1, device_mesh, sharding)
-        batch_2_dt = distribute_tensor(batch_2, device_mesh, sharding)
-        new_dt = torch.baddbmm(
-            tensor_dt, batch_1_dt, batch_2_dt, beta=0.0, alpha=0.5
-        )
-        self.assertTrue(tensor.requires_grad)
-        self.assertTrue(tensor_dt.requires_grad)
-        self.assertTrue(new_dt.placements[0].is_shard(dim=0))
-        local_result.sum().backward()
-        new_dt.sum().backward()
-        batch_1_dt_grad = torch.empty(
-            batch_1_dt.grad.size(),
-            dtype=batch_1_dt.grad.dtype,
-            device=self.device_type,
-        )
-        dist._all_gather_base(batch_1_dt_grad, batch_1_dt.grad.to_local())
-        self.assertEqual(batch_1_dt_grad, batch_1.grad)
-        self.assertEqual(tensor_dt.grad.to_local(), tensor.grad)
-
-    @with_comms
-    def test_sharded_bmm(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        torch.manual_seed(self.rank)
-        input = torch.rand(3, 5, 8, device=self.device_type)
-        mat_2 = torch.rand(3, 8, 6, device=self.device_type)
-        local_result = torch.bmm(input, mat_2)
-        sharding = [Shard(0)]
-        input_dt = DTensor.from_local(input, device_mesh, sharding)
-        mat_2_dt = DTensor.from_local(mat_2, device_mesh, sharding)
-        new_dt = torch.bmm(input_dt, mat_2_dt)
-        self.assertTrue(new_dt.placements[0].is_shard(dim=0))
-        self.assertEqual(new_dt.to_local(), local_result)
-
-    @with_comms
-    def test_sharded_bmm_with_bwd(self):
-        device_mesh = DeviceMesh(self.device_type, list(range(self.world_size)))
-        torch.manual_seed(10)
-        input = torch.rand(
-            16, 5, 8, device=self.device_type, requires_grad=True
-        )
-        mat_2 = torch.rand(
-            16, 8, 6, device=self.device_type, requires_grad=True
-        )
-        local_result = torch.bmm(input, mat_2)
-        sharding = [Shard(0)]
-        input_dt = distribute_tensor(input, device_mesh, sharding)
-        mat_2_dt = distribute_tensor(mat_2, device_mesh, sharding)
-        new_dt = torch.bmm(input_dt, mat_2_dt)
-        self.assertTrue(input.requires_grad)
-        self.assertTrue(input_dt.requires_grad)
-        self.assertTrue(new_dt.placements[0].is_shard(dim=0))
-        local_result.sum().backward()
-        new_dt.sum().backward()
-        input_dt_grad = torch.empty(
-            input_dt.grad.size(),
-            dtype=input_dt.grad.dtype,
-            device=self.device_type,
-        )
-        dist._all_gather_base(input_dt_grad, input_dt.grad.to_local())
-        self.assertEqual(input_dt_grad, input.grad)
-        mat_2_dt_grad = torch.empty(
-            mat_2_dt.grad.size(),
-            dtype=mat_2_dt.grad.dtype,
-            device=self.device_type,
-        )
-        dist._all_gather_base(mat_2_dt_grad, mat_2_dt.grad.to_local())
-        self.assertEqual(mat_2_dt_grad, mat_2.grad)
-
     @with_comms
     def test_sharded_softmax(self):
         for softmax_dim in [1, 2, -1]:
